feature: add authE to comment

+++ backend/src/main/java/com/revature/backend/comments/CommentService.java
@@ -1,10 +1,12 @@
 package com.revature.backend.comments;
 
+import com.revature.backend.auth.InvalidCredentialsException;
 import com.revature.backend.posts.Post;
 import com.revature.backend.posts.PostRepository;
 import com.revature.backend.users.User;
 import com.revature.backend.users.UserRepository;
 import com.revature.backend.utils.EntityNotFoundException;
+import com.revature.backend.utils.IAuthenticationFacade;
 import jakarta.validation.Valid;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,10 +22,30 @@ public class CommentService {
 	private final UserRepository userRepository;
 	private final PostRepository postRepository;
 	private final CommentRepository commentRepository;
+	private final IAuthenticationFacade authenticationFacade;
+
+	private void checkCommentOwnership(Comment comment) {
+		/*
+		 * Check if the logged-in user can modify the comment specified by the URL path.
+		 * TODO: add role-based check; if logged-in as admin, should be able to bypass ownership check.
+		 */
+		Integer claimedUserId = authenticationFacade.getClaimedUserId();
+
+		Optional<User> claimedUser = userRepository.findByIdAndDeletedFalse(claimedUserId);
+		if (claimedUser.isEmpty()) {
+			throw new InvalidCredentialsException(
+					String.format(
+							"Current user %s is deleted",
+							claimedUserId));
+		}
+
+		if (!claimedUserId.equals(comment.getAuthor().getId())) {
+			throw new InvalidCredentialsException(claimedUserId, "comment", comment.getId());
+		}
+	}
 
 	public Comment createComment(@Valid CommentDto commentDto) {
-		// TODO: auth, author id, etc
-		Integer authorId = commentDto.getAuthorId();
+		Integer authorId = authenticationFacade.getClaimedUserId();
 		Integer postId = commentDto.getPostId();
 		Optional<User> author = userRepository.findByIdAndDeletedFalse(authorId);
 		Optional<Post> post = postRepository.findById(postId);
@@ -56,6 +78,8 @@ public class CommentService {
 		}
 
 		Comment c = comment.get();
+		checkCommentOwnership(c);
+
 		c.setUpdatedAt(LocalDateTime.now());
 		c.setContent(commentDto.getContent());
 		return commentRepository.save(c);
@@ -66,6 +90,9 @@ public class CommentService {
 		if (comment.isEmpty()) {
 			throw new EntityNotFoundException("comment", commentId);
 		}
+
+		Comment c = comment.get();
+		checkCommentOwnership(c);
 
 		commentRepository.deleteById(commentId);
 	}
